+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - New method `PayloadInterface::getOrFail(string $path): mixed`
 - Support of wildcard "*" at the end of job name in `JobHandlerManager`
+- Support of wildcard "*" at the end of queue name in `QueueManager::filter()` method
 
 ## [1.0.0-beta4] - 2025-01-20